+++ mi-train/DBuz747/fxmanifest.lua
@@ -1,0 +1,12 @@
+fx_version 'cerulean'
+game 'gta5'
+
+name 'DBuz747'
+description 'DBuz747 passenger wagon stream for jp-mi-train (MrGTAmodsgerman add-on)'
+version '1.1.0'
+
+files {
+    'data/vehicles.meta',
+}
+
+data_file 'VEHICLE_METADATA_FILE' 'data/vehicles.meta'